Fix to compare against signed/unsigned warning
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@3200 e03df62e-2008-0410-955e-edbf42e46eb7

+++ src/socket.cpp
@@ -214,7 +214,7 @@ char* InspSocket::Read()
 	if ((fd < 0) || (fd > MAX_DESCRIPTORS))
 		return NULL;
 	int n = recv(this->fd,this->ibuf,sizeof(this->ibuf),0);
-	if ((n > 0) && (n <= sizeof(this->ibuf)))
+	if ((n > 0) && (n <= (int)sizeof(this->ibuf)))
 	{
 		ibuf[n] = 0;
 		return ibuf;
